modularice quetions
