+++ src/types/general.types.d.ts
@@ -1,0 +1,7 @@
+
+export type QuestionType = {
+    question: string
+    solution: string
+    extra?: string
+  }
+
